Added COI structure to networkclient
